pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "word-ast"
version = "0.1.0"
description = "Word <-> AST round-trip toolkit"
requires-python = ">=3.10"
dependencies = [
  "python-docx>=1.1.0",
  "lxml>=5.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
